small fixes + ReadMe updated both v02+ HFL25-2
