+++ README.md
@@ -4,4 +4,9 @@ Hybrid development with Flutter - Course
 ## v01
 Simple calculator CLI that uses `+ - / *` and gives a result in the `double` datatype.  
 
-> ![Picture of CLI UI](/assets/img/KalkylatorImg.JPG)
+> ![Picture of CLI UI Calculator](/assets/img/KalkylatorImg.JPG)
+
+## v02
+Simple CLI UI for handling all the heroes that one could need
+
+> ![Picture of CLI UI HeroDex3000](/assets/img/HeroDex3000Menu.JPG)
