+++ v02/README.md
@@ -1,2 +1,10 @@
-# v02
-Automatically generated README for v02
+# Simple Dart project - CLI HeroDex3000
+
+*A simple command-line application with an entrypoint in `bin/`, library code in `lib/`, and example unit tests in `test/`.*
+
+## How to run the program
+1. Open a terminal.
+2. Navigate to the `v02` folder.
+3. Run:
+   `dart run bin/v02.dart`
+4. Follow intructions to `add hero`, `show all heroes`. `search for a hero`.
