+++ v02/lib/utils/console_utils.dart
@@ -45,6 +45,7 @@ class ConsoleUtils {
     // }
   }
 
+  // Function to give message to user about invalid input
   void invalidChoice(){
     stdout.writeln("Ogiltigt input. Vänligen försök igen");
     sleep(Duration(seconds: 2));
