+++ v02/lib/utils/menu.dart
@@ -1,3 +1,5 @@
+
+// Class to create a customizable menu in a CLI with multiple options
 class Menu{
 
   final List<String> _options;
