--- v02/test/v02_test.dart
@@ -1,8 +1,0 @@
-// import 'package:v02/v02.dart';
-// import 'package:test/test.dart';
-
-// void main() {
-//   test('calculate', () {
-//     expect(calculate(), 42);
-//   });
-// }
